Review feedback and more fixes.
